feat: convert to typescript
- Only named exports
- No more CJS, only ESM
- Adds auto-publish
- Adds dependabot

BREAKING CHANGE: switch to named exports, ESM only
