+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules
+dist
+docs
+package-lock.json
 .nyc_output
-coverage
