--- .npmignore
@@ -1,4 +1,0 @@
-node_modules
-.nyc_output
-coverage
-test.js
