+++ src/index.ts
@@ -1,8 +1,14 @@
+
 // If the passed object is an (async) iterable, then get the iterator
 // If it's probably an iterator already (i.e. has next function) return it
 // else throw
-module.exports = function getIterator (obj) {
-  if (obj) {
+export function getIterator <T> (obj: AsyncIterable<T>): AsyncIterator<T>
+export function getIterator <T> (obj: AsyncIterator<T>): AsyncIterator<T>
+export function getIterator <T> (obj: Iterable<T>): Iterator<T>
+export function getIterator <T> (obj: Iterator<T>): Iterator<T>
+export function getIterator <T> (obj: any): AsyncIterator<T> | Iterator <T>
+export function getIterator <T> (obj: any): AsyncIterator<T> | Iterator <T> {
+  if (obj != null) {
     if (typeof obj[Symbol.iterator] === 'function') {
       return obj[Symbol.iterator]()
     }
